Norwegian navy and army overhaul
Norway started construction of Paddle Steamers and planned a Battleship doctrine line in 1833. Drammen was the main base of the navy and was where the construction of paddle steamers started. There was also a military shipyard in Trondhjem. General "Johan Henrik Spørck" got into a military academy aged frickin' 11 and he had had huge talent in leading troops. He was well educated and well experienced. In 1836, there were 8000 guards in Norway who were recruited because the Norwegians tried to build up against another Swedish attack to fully annex Norway. (Swedes got very insulted about the Norwegian implemention of 3 guard brigades). Admiral Thomas Konow became a sea officer at the age of 17 when the minimum age limit actually was 25. He is remembered as one of the most new-thinking admirals of his time and a pioner.