Sardu (explanation on the desc)
The Sardinian people, which you added, represent the Sardu, a Romance language spoken by the Sardinians on the Western Mediterranean island of Sardinia, with large influences from the Catalan from l'Alghero (L'Alguer), and name their language as "Sardu". If you want to represent the language in English, it would be "Sard".