complex DN thing, took long (read desc)
Ok, so basically originally there was a problem in DN with countries using the republic_5, republic_6, republic_7, etc. governments releasing countries and them using the blue flag. Bob fixed this by giving every country flag redirects to the theocracy slot if they have the _5th/_6th/_7th/... government. However, for many countries the theocracy slot uses an atypical flag, for example the Belgian flag in the theocracy slot is for their weird vassaldom under Germany or for example several new world countries have the spanish vassal flag in their theocracy slot. In all, it'd be less confusing to use democracy as the default for the _5th/_6th/_7th/... governments, which is why I implemented it.