READ DESC
For some reason, these pops promoted to intellectuals/bureaucrats IF their nation was an AI and had 74% of admin/education spending. This was for AI only on a separate modifier, i removed the separate modifier and the AI requirement, now it works for players too, and it's technically a performance boost.