bulgarian polygyros properly implemented
pauil added OR bulgarian to the random_owned of the Greek renaming for Polygyros, however this is incorrect in two regards: Firstly, he neglected to do it in the allow, so it wouldn't work anyway. Secondly, the Bulgarian name for Polygyros is Poligiros, so the renaming wasn't even correct. Anyway, I added it properly.