Please don't undo my Sino-Korean renamings plz
And also, there's no need to add "-si" or "-shi" as a suffix to any of these province names in Korean or Japanese. Officially, almost every city in Korea and Japan is referred to with the -shi suffix, meaning city, in its name. But no other provinces in this game make use of that; they use their more common names, which drop the suffix.