Iran doesnt accept Lurish anymore + Arabs in Luristan and Lurs in Qazvin
Reading about the anarchical state of Luristan during the Qajar period along with the harsh if not outright hostile relations between Lurish tribes and the central government has lead me to conclude that having them as an accepted culture is wholly inappropriate.
The only one of these cultures that I would consider even remotely appropriate would be the Bakhtiari, which, while also a Lurish tribe, dont seem to have underwent the same level of treatment as their northern cousins.
It remains to be seen if further reading will change my mind on this, but in order to be on the safe side, I will also remove them from the PER accepted cultures list until I finish my research.